Add first pass file io
- Fasta file, including multi-line fasta
- FastQ file, no affordance for multi-line (the ambiguity would be a
  little much, shouldn't be a problem though)
- SnapGene file, simple approach

The over arching goal is to have the parsers be as simple as possible.
They read only the data they expect, and they return the most minimal
data strcture that makes sense.

In the case of Fasta and FastQ, that means a list of binaries. These are
implicitly ordered, and returned in reverse because of the recursive
approach.

For the SnapGene files, things are a little more complicated. SnapGene's
format include XML portions. In order to expose that simply, I opted to
just return the `:xmerl` data structure. This would allow anyone to use
an x-path into the data and select what they need.

This is all predicated on the idea that _I have no idea what I'm doing_.