Add basic polymer type with standard sequences
This adds the basics of a polymer and sequence types reflecting normal
biological molecules. There's a lot to be desired here, but this is the
basic structure that I think I'll try and role with into v1.

The idea is that the `Bio.Polymer` forms the basis for all polymers and
will likely eventually be a behavior. I just need to sort out how to
have concrete implementations that are then overridden in other modules.
Good learning experience!

The sequences themselves are something that I want users to interact
with comfortably, meaning that `Enumerable` really needed to be
implemented for all of them. I went with `using` the `Bio.Sequence`
module for this as it simplified sharing of that code/logic.

I really do love doctests.