Add Utilities.slide/3
Because the implementation for `slide/3` in the Enum module behaves the
way it does, we should just let it. In fact, I'm considering if I
shouldn't just use it as a basis for creating an internal module?

That is, I'm not going to re-implement the Enum module and Enumerable
protocol. Instead, I would just have `Bio.Enum` and expose all the same
functions but with the explicit guarantee that they preserve the struct
type that you're using.

This would alleviate any irritation from the perspective of the behavior
of e.g. `Enum.map` returning a struct instead of a list. You would have
to explicitly opt in to one behavior or the other...

Yeah, I kinda really like that.