SnapGene parses to struct
Both the module itself, and the sequence that it parses are now returned
as structs based on the sequence. This provides some utility in
determining the type of the data, which I believe may be allowed to be
more than just a DNA string? We'll have to see about that. This leaves
doors open without too much trouble though.