[feature] Update sequences for easier conversion
After working with the library for a little while solving some problems
on Project Rosalind, I determined that the mechanics for defining the
conversions weren't right. Namely, requiring the definition of a
protocol was overly burdensome.

That lead me to thinking about how to generalize the nature of the
thing. That kicked off the idea of k-mer enumeration. That is, the
enumerable properties of the sequences are well defined. We want to have
those be semantically similar to other languages like Python. However,
for conversions, we may need to enumerate k-mers!

So instead of defining a protocol for doing the conversion wholesale, we
keep the Polymer module hooked up to read from a module, but we let that
function receive the results of a k-mer enumeration! That way the
conversion defines the k-mer size as well as how to go from element to
element.

This feels much better to me.

The generalization that mad sense then was to say that these are the
`Polymeric` types. This also cemented where I could put the `valid?` and
`validate` functions to allow for polymer validations!